[変更]プレイリストの依存関係を改善した
