+++ Niconicome/Models/Domain/Local/Store/PlaylistStoreHandler.cs
@@ -5,6 +5,7 @@ using Niconicome.Models.Playlist;
 using STypes = Niconicome.Models.Domain.Local.Store.Types;
 using Niconicome.Extensions.System.List;
 using Niconicome.Models.Helper.Result;
+using Niconicome.Models.Playlist.Playlist;
 
 namespace Niconicome.Models.Domain.Local.Store
 {
