+++ Niconicome/ViewModels/Mainpage/PlayistTreeViewModel.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Diagnostics;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
@@ -10,11 +9,11 @@ using System.Windows.Media;
 using Microsoft.Xaml.Behaviors;
 using Niconicome.Extensions;
 using Niconicome.Extensions.System.Windows;
-using Niconicome.Models.Playlist;
+using Niconicome.Models.Playlist.Playlist;
 using Niconicome.ViewModels.Controls;
 using Niconicome.ViewModels.Mainpage.Subwindows;
 using Niconicome.Views;
-using Utils = Niconicome.Models.Domain.Utils;
+using Reactive.Bindings;
 using WS = Niconicome.Workspaces;
 
 namespace Niconicome.ViewModels.Mainpage
@@ -30,29 +29,13 @@ namespace Niconicome.ViewModels.Mainpage
         public PlaylistTreeViewModel(Func<string, MessageBoxButtons, MessageBoxIcons, Task<MaterialMessageBoxResult>> showMessageBox)
         {
             this.ShowMessageBox = showMessageBox;
+            this.PlaylistTree = WS::Mainpage.PlaylistTreeHandler.Playlists.ToReadOnlyReactiveCollection(x => x);
 
-            //初期化失敗時
-            void initializeFailed()
-            {
-                this.ShowMessageBox("アプリケーションの初期化中にエラーが発生しました。詳しくはエラーログをご確認下さい。", MessageBoxButtons.OK, MessageBoxIcons.Error);
-                Application.Current?.Shutdown();
-            }
-
-            try
-            {
-                this.PlaylistTreeHandler = WS::Mainpage.PlaylistTree;
-            }
-            catch (Exception e)
-            {
-                Utils::Logger.GetLogger().Error("初期化中にエラーが発生しました。", e);
-                initializeFailed();
-            }
 
 #pragma warning disable CS8622
             this.AddPlaylist = new CommandBase<int>((object? arg) => true, (int parent) =>
             {
-                WS::Mainpage.PlaylistTree.AddPlaylist(parent);
-                this.OnPropertyChanged(nameof(this.PlaylistTree));
+                WS::Mainpage.PlaylistHandler.AddPlaylist(parent);
             });
             this.RemovePlaylist = new CommandBase<int>((object? arg) => true, async (int playlist) =>
             {
@@ -60,19 +43,17 @@ namespace Niconicome.ViewModels.Mainpage
 
                 if (result != MaterialMessageBoxResult.Yes) return;
 
-                WS::Mainpage.PlaylistTree.DeletePlaylist(playlist);
-                this.OnPropertyChanged(nameof(this.PlaylistTree));
+                WS::Mainpage.PlaylistHandler.DeletePlaylist(playlist);
             });
 #pragma warning restore CS8622
 
             this.PlaylistRefreshcommand = new CommandBase<object>(_ => true, _ =>
             {
-                WS::Mainpage.PlaylistTree.Refresh();
+                WS::Mainpage.PlaylistHandler.Refresh();
                 if (WS::Mainpage.CurrentPlaylist.SelectedPlaylist.Value is not null)
                 {
-                    WS::Mainpage.CurrentPlaylist.SelectedPlaylist.Value = WS::Mainpage.PlaylistTree.GetPlaylist(WS::Mainpage.CurrentPlaylist.SelectedPlaylist.Value.Id);
+                    WS::Mainpage.CurrentPlaylist.SelectedPlaylist.Value = WS::Mainpage.PlaylistHandler.GetPlaylist(WS::Mainpage.CurrentPlaylist.SelectedPlaylist.Value.Id);
                 }
-                this.OnPropertyChanged(nameof(this.PlaylistTree));
             });
 
             this.EditPlaylistCommand = new CommandBase<object>(_ => true, arg =>
@@ -97,21 +78,11 @@ namespace Niconicome.ViewModels.Mainpage
         /// </summary>
         private Func<string, MessageBoxButtons, MessageBoxIcons, Task<MaterialMessageBoxResult>> ShowMessageBox { get; set; }
 
-        /// <summary>
-        /// プレイリストのツリー
-        /// </summary>
-        public IPlaylistHandler? PlaylistTreeHandler { get; private set; }
 
         /// <summary>
         /// プレイリストのツリー
         /// </summary>
-        public ObservableCollection<ITreePlaylistInfo> PlaylistTree
-        {
-            get
-            {
-                return this.PlaylistTreeHandler?.Playlists ?? new ObservableCollection<ITreePlaylistInfo>();
-            }
-        }
+        public ReadOnlyReactiveCollection<ITreePlaylistInfo> PlaylistTree { get; init; }
 
         /// <summary>
         /// プレイリスト追加
@@ -346,7 +317,7 @@ namespace Niconicome.ViewModels.Mainpage
 
             if (this.insertType == InsertType.Child)
             {
-                context.PlaylistTreeHandler?.Move(sourceInfo.Id, parentInfo.Id);
+                WS::Mainpage.PlaylistHandler.Move(sourceInfo.Id, parentInfo.Id);
                 this.insertType = InsertType.None;
             }
         }
