+++ NiconicomeTest/Local/PlaylistInfoUnitTest.cs
@@ -6,6 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Niconicome.Models.Playlist;
 using STypes = Niconicome.Models.Domain.Local.Store.Types;
+using Niconicome.Models.Playlist.Playlist;
 
 namespace NiconicomeTest.Local.Playlist
 {
@@ -43,7 +44,7 @@ namespace NiconicomeTest.Local.Playlist
     [TestFixture]
     class PlaylistInfoHandlerUnitTest
     {
-        private IPlaylistTreeConstructor handler = new PlaylistTreeConstructor();
+        private IPlaylistTreeHandler handler = new PlaylistTreeHandler();
 
         [SetUp]
         public void SetUp()
@@ -72,8 +73,8 @@ namespace NiconicomeTest.Local.Playlist
 
             //全て追加
             var playlists = new List<STypes::Playlist>() { first, second, third, fourth, fifth, sixth, seventh };
-            this.handler = new PlaylistTreeConstructor();
-            this.handler.AddRange(playlists.Select(p => NonBindableTreePlaylistInfo.ConvertToTreePlaylistInfo(p, playlists.Where(pl => pl?.ParentPlaylist?.Id == p.Id))));
+            this.handler = new PlaylistTreeHandler();
+            this.handler.Initialize(playlists.Select(p => NonBindableTreePlaylistInfo.ConvertToTreePlaylistInfo(p, playlists.Where(pl => pl?.ParentPlaylist?.Id == p.Id))));
         }
 
         [TestCase(1)]
@@ -88,12 +89,6 @@ namespace NiconicomeTest.Local.Playlist
             Assert.IsTrue(this.handler.Contains(id));
         }
 
-        [Test]
-        public void プレイリストを追加する()
-        {
-            this.handler.Add(new NonBindableTreePlaylistInfo() { Id = 8 });
-            Assert.IsTrue(this.handler.Contains(8));
-        }
 
         [Test]
         public void プレイリストを削除する()
@@ -127,17 +122,11 @@ namespace NiconicomeTest.Local.Playlist
         }
 
         [Test]
-        public void ルートプレイリストを取得する()
-        {
-            ITreePlaylistInfo root = this.handler.GetRoot();
-            Assert.AreEqual(1, root.Id);
-        }
-
-        [Test]
         public void ツリーを構築する()
         {
-            ITreePlaylistInfo tree = this.handler.GetTree();
+            ITreePlaylistInfo tree = this.handler.Playlists.FirstOrDefault()!;
 
+            Assert.That(tree, Is.Not.Null);
             //子プレイリストの概観チェック
             Assert.AreEqual(2, tree.Children.Count);
             Assert.AreEqual(1, tree.Children[0].Children.Count);
