+++ NiconicomeTest/Stabs/Models/Domain/Local/Store/PlaylistStoreHandlerStab.cs
@@ -4,7 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Niconicome.Models.Domain.Local.Store;
+using Niconicome.Models.Helper.Result;
 using Niconicome.Models.Playlist;
+using Niconicome.Models.Playlist.Playlist;
 using STypes = Niconicome.Models.Domain.Local.Store.Types;
 
 namespace NiconicomeTest.Stabs.Models.Domain.Local.Store
@@ -58,6 +60,18 @@ namespace NiconicomeTest.Stabs.Models.Domain.Local.Store
         public IEnumerable<STypes::Playlist> GetAllPlaylists()
         {
             return new List<STypes::Playlist>() { this.GetRootPlaylist() };
+        }
+
+
+        public IAttemptResult MoveVideoToPrev(int playlistID, int videoIndex)
+        {
+            return new AttemptResult() { IsSucceeded = true };
+        }
+
+        public IAttemptResult MoveVideoToForward(int playlistID, int videoIndex)
+        {
+            return new AttemptResult() { IsSucceeded = true };
+
         }
 
         public void RemoveChildPlaylist(int selfId)
